Add route, controller and views for adding, editing and viewing manufacturers
Deleting manufacturers has been ignored for the
time being. We will want to avoid deleting a
manufacturer if there is a product using them,
or an order that contains a product that
uses the manufacturer. Too complicated to do with
code, but probably easy to do with SQL when a
proper database is set up.