Move uploading of images to separate form
Just for simplicity's sake really - otherwise,
the product form would actually be submitting data
for 2 tables on the database: a product and one or
more product images. The product images have a se-
parate table as you should be able to upload many
photos for one product.

So now there is a form for creating/editing produ-
cts, and a different one for uploading product ph-
otos. It is not intuitive or a good user experience,
but for the time being I want to finish this project
and not spend hours trying to figure out how to
upload photos to one data table, and other info to
another table, all as part of one form.

Perhaps I should have had the initial product form,
and then a link on each product to a separate form
for uploading photos. I will change this later,
as it is a more inuitive way of doing it than what
I have now.